Add vello as a rendering backend
Also does a bit of `Renderer` refactoring:
* Move size and `Context` out of `Renderer` and pass them in as arguments where needed
* Introduce a `Scene` type that is exposed to users so that they are shielded from the `Renderer` type

I've also added another top-level `Scene` type that is just an enum of `renderer::Scene` and `vello::Scene` and its value is based on whichever rendering backend was chosen when constructing the game engine.

That enum should probably be a struct so that user can construct and combine all the scenes at once. And egui scene (its `Ui` type) should also be exposed in a similar fashion. But more thought is needed for how to combine all these scenes and backends correctly.